requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiosqlite>=0.19.0
plaid-python>=15.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
